refactor(coordinator): relocate next-gen coordinator into unified::GroupCoordinator (64d-B B3)
